pyproject.toml
[tool.poetry]
name = "socialmedia"
version = "0.1.0"
description = ""
authors = ["Mohammed Salah <mohamed.salah9972@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
django-model-utils = "^4.3.1"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
